Changes to interrupt handling to fix Digger sound glitches.
Problem was that interrupt requests would be lost because
Digger is almost always in interrupt disabled state. Solution
was that the CPU and daisy chain no longer cancels pending interrupt
requests, but keeps the interrupt request alive until it can be
serviced. On real hardware, the request would 'expire' at some point,
this is currently not implemented.